requirement.txt
annotated-types==0.7.0
ansible-core==2.16.0
anyio==4.10.0
asyncio==4.0.0
certifi==2025.8.3
cffi==1.17.1
cryptography==45.0.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
Markdown==3.8.2
MarkupSafe==3.0.2
micromlgen==1.1.28
numpy==2.3.2
openai==1.100.1
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.2
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
pymdown-extensions==10.16.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
resolvelib==1.0.1
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
websockets==15.0.1
